requirements.txt
Django>=1.5
djangorestframework>=2.3
Markdown>=2.1
django-filter>=0.5.4
django-oauth2-provider>=0.2.3
django-guardian>=1.1.1
Pygments
Pillow
pep8
django-imagekit
django-bootstrap3
